keylogger/requirements.txt
pynput
pyperclip
py-machineid
pillow
Crypto